requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
langchain
langchain_ollama
langchain-community
duckduckgo-search
ddgs
weaviate-client